talkback_demo/requirements.txt
Flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
eventlet==0.33.3
paho-mqtt==1.6.1
Werkzeug==3.0.1
